Move SignStatusWrapper from Sign.tsx into SignStatus.tsx. Import colors from brand repo there and tweak some styling.
